Simplified pokemon database and repository structure, splitting them into shell and core respectively
