--- src/core/repositories/mod.rs
@@ -1,1 +1,0 @@
-pub mod pokedex;
