+++ src/shell/mod.rs
@@ -1,2 +1,2 @@
-pub mod database;
+pub mod pokemon;
 pub mod ui;
